+++ CHANGELOG.md
@@ -6,6 +6,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## [2.0.0] - 2019-12-03
+## Breaking changes
+- **Bump PHP requirement >= 7.3**.
+- **Bump WordPress requirement >= 5.1**.
+### Updates
+- Update the PHP and WordPress versions in Travis.
+- Remove the PHPMD library.
+- Update the gitattributes to ignore /vendor and itself.
+- Update minimum PHP version, upgrade symfony/http-foundation to version 4.2.12 or later or version 5.
+- Bump PHP Unit and WP PHP Unit.
+- Remove PHPMD in composer scripts and bin/.
+- Adding new Option class for settings integrations.
+- Cleanup of the Service Provider since a factory isn't needed in a DI model.
+### Changes
+- All tasks need to be their `clearScheduledEvent` method to be updated, 
+see [ExampleMigrationTask.php](src/Tasks/ExampleMigrationTask.php), as a second param is now required to pass in
+the `UpgradeModel`, which is injected into the `dispatch` method.
+
 ## [1.3.0] - 2019-11-25
 - Make sure WordPress' option for `timezone_string` returns a string, update to use helper method which will call
 WordPress's new `wp_timezone_string` if available (WordPress >= 5.3) otherwise uses it's code. Fixes #9.